First working version of "Tweet this" add-on.
The add-on uses the text of the rendered item as text of the tweet
together with the item's URL. The add-on injects Twitter's script
once and call twttr.widgets.load() whenever a button needs to be
rendered. That works although that's not documented anywhere on
Twitter's Web site as far as I can tell.

Next updates to do on the add-on:
- define a couple of configuration parameters to adjust the button
and the message to be tweeted
- improve support for items. There may be cases where the URL does
not mean anything or where we should fallback to e.g. contentURL.